Change sensor.timestamp type to fit ms since epoch.

+++ phoenix_server/priv/repo/migrations/20151129144507_create_sensor.exs
@@ -3,7 +3,7 @@ defmodule PhoenixServer.Repo.Migrations.CreateSensor do
 
   def change do
     create table(:sensors) do
-      add :timestamp, :integer
+      add :timestamp, :bigint
       add :device, :string
       add :x, :float
       add :y, :float
